refactor(numerical-semigroups): isolate factorization contracts
