+++ src/jacobian/math/numerical_semigroups/_models.py
@@ -247,59 +247,6 @@ def _edges_span(component_count: int, edges: list[tuple[int, int]]) -> bool:
         reached = expanded
 
 
-def _require_exact_factorization_family(
-    generators: tuple[int, ...],
-    value: int,
-    family: tuple[tuple[int, ...], ...],
-) -> None:
-    if len(set(family)) != len(family):
-        raise _validation_error("factorizations must be unique")
-    for factorization in family:
-        if len(factorization) != len(generators) or any(
-            coordinate < 0 for coordinate in factorization
-        ):
-            raise _validation_error("factorization has invalid coordinates")
-        degree = sum(
-            coordinate * generator
-            for coordinate, generator in zip(factorization, generators, strict=True)
-        )
-        if degree != value:
-            raise _validation_error(
-                "factorization does not evaluate to the result value"
-            )
-    if len(family) != factorization_count(generators, value):
-        raise _validation_error("factorizations do not form the complete family")
-
-
-def _factorization_graph_data(
-    family: tuple[tuple[int, ...], ...],
-) -> tuple[tuple[tuple[int, int], ...], tuple[tuple[int, ...], ...]]:
-    edges = tuple(
-        (left, right)
-        for left in range(len(family))
-        for right in range(left + 1, len(family))
-        if any(
-            a > 0 and b > 0 for a, b in zip(family[left], family[right], strict=True)
-        )
-    )
-    unseen = set(range(len(family)))
-    components: list[tuple[int, ...]] = []
-    while unseen:
-        reached = {min(unseen)}
-        while True:
-            expanded = reached | {
-                right if left in reached else left
-                for left, right in edges
-                if left in reached or right in reached
-            }
-            if expanded == reached:
-                break
-            reached = expanded
-        unseen.difference_update(reached)
-        components.append(tuple(sorted(reached)))
-    return edges, tuple(components)
-
-
 class NumericalSemigroupRequest(StrictModel):
     """A numerical semigroup defined by a finite set of positive generators."""
 
@@ -310,225 +257,6 @@ class NumericalSemigroupRequest(StrictModel):
     @model_validator(mode="after")
     def require_positive_generators(self) -> Self:
         _require_positive_bounded_generators(self.generators)
-        return self
-
-
-# ---------------------------------------------------------------------------
-# Extended operations: factorization, elasticity, catenary degree, etc.
-# ---------------------------------------------------------------------------
-
-
-class FactorizationComputeRequest(StrictModel):
-    """Compute the complete factorization family Z(s) for one element."""
-
-    generators: tuple[CanonicalInteger, ...] = Field(
-        min_length=1,
-        max_length=MAX_GENERATORS,
-        description=(
-            f"Positive generators with gcd 1, each at most {MAX_GENERATOR}. "
-            "The presentation may be reordered or redundant; returned "
-            "coordinates use its increasing minimal generator axis."
-        ),
-    )
-    value: CanonicalInteger
-
-    @model_validator(mode="after")
-    def require_complete_materialization(self) -> Self:
-        generators = _require_minimal_generators(self.generators)
-        value = _require_bounded_value(self.value)
-        _require_materializable_factorizations(
-            generators, value, MAX_MATERIALIZED_FACTORIZATIONS
-        )
-        return self
-
-
-class FactorizationComputeResult(StrictModel):
-    """Complete factorization family Z(s) for one element."""
-
-    value: CanonicalInteger
-    minimal_generators: tuple[CanonicalInteger, ...] = Field(
-        min_length=1, max_length=MAX_GENERATORS
-    )
-    in_semigroup: bool
-    factorizations: tuple[tuple[int, ...], ...]
-
-    @model_validator(mode="after")
-    def require_exact_family(self) -> Self:
-        generators = _require_canonical_minimal_axis(self.minimal_generators)
-        value = parse_canonical_integer(self.value)
-        if self.in_semigroup != bool(self.factorizations):
-            raise _validation_error(
-                "membership must agree with the factorization family"
-            )
-        _require_exact_factorization_family(generators, value, self.factorizations)
-        return self
-
-
-class FactorizationLengthsComputeRequest(StrictModel):
-    """Compute the complete sorted length set of one element."""
-
-    generators: tuple[CanonicalInteger, ...] = Field(
-        min_length=1,
-        max_length=MAX_GENERATORS,
-        description=(
-            f"Positive generators with gcd 1, each at most {MAX_GENERATOR}. "
-            "The presentation may be reordered or redundant; factorization "
-            "lengths use its increasing minimal generator axis."
-        ),
-    )
-    value: CanonicalInteger
-
-    @model_validator(mode="after")
-    def require_positive_generators_and_bounded_value(self) -> Self:
-        _require_minimal_generators(self.generators)
-        _require_bounded_value(self.value)
-        return self
-
-
-class FactorizationLengthsComputeResult(StrictModel):
-    """Sorted length set of one element."""
-
-    value: CanonicalInteger
-    minimal_generators: tuple[CanonicalInteger, ...] = Field(
-        min_length=1, max_length=MAX_GENERATORS
-    )
-    in_semigroup: bool
-    lengths: tuple[int, ...]
-
-    @model_validator(mode="after")
-    def require_length_set(self) -> Self:
-        generators = _require_canonical_minimal_axis(self.minimal_generators)
-        value = parse_canonical_integer(self.value)
-        if self.in_semigroup != bool(self.lengths):
-            raise _validation_error(
-                "membership must agree with the factorization lengths"
-            )
-        if self.lengths != tuple(sorted(set(self.lengths))):
-            raise _validation_error(
-                "lengths must be strictly increasing and duplicate-free"
-            )
-        if any(length < 0 for length in self.lengths):
-            raise _validation_error("factorization lengths must be non-negative")
-        if self.lengths != factorization_lengths(generators, value):
-            raise _validation_error("lengths do not form the complete length set")
-        return self
-
-
-class FactorizationDistanceRequest(StrictModel):
-    """Distance between two factorizations of the same element."""
-
-    generators: tuple[CanonicalInteger, ...] = Field(
-        min_length=1,
-        max_length=MAX_GENERATORS,
-        description=(
-            f"Positive generators with gcd 1, each at most {MAX_GENERATOR}. "
-            "The presentation may be reordered or redundant, but both "
-            "factorization coordinate tuples must use its increasing minimal "
-            "generator axis."
-        ),
-    )
-    value: CanonicalInteger
-    first: tuple[int, ...] = Field(min_length=1)
-    second: tuple[int, ...] = Field(min_length=1)
-
-    @model_validator(mode="after")
-    def require_valid_request(self) -> Self:
-        generators = _require_minimal_generators(self.generators)
-        value = _require_bounded_value(self.value)
-        if any(c < 0 for c in self.first) or any(c < 0 for c in self.second):
-            raise _validation_error("factorization coordinates must be non-negative")
-        if len(self.first) != len(generators) or len(self.second) != len(generators):
-            raise _validation_error(
-                "factorization coordinates must match the minimal generating system"
-            )
-        first_value = sum(
-            coefficient * generator
-            for coefficient, generator in zip(self.first, generators, strict=True)
-        )
-        second_value = sum(
-            coefficient * generator
-            for coefficient, generator in zip(self.second, generators, strict=True)
-        )
-        if first_value != value or second_value != value:
-            raise _validation_error(
-                "both factorizations must evaluate to the declared value"
-            )
-        return self
-
-
-class FactorizationDistanceResult(StrictModel):
-    """Distance between two factorizations."""
-
-    value: CanonicalInteger
-    distance: int = Field(ge=0)
-    first_length: int = Field(ge=0)
-    second_length: int = Field(ge=0)
-
-
-class FactorizationGraphComputeRequest(StrictModel):
-    """Compute the standard factorization graph of one element."""
-
-    generators: tuple[CanonicalInteger, ...] = Field(
-        min_length=1,
-        max_length=MAX_GENERATORS,
-        description=(
-            f"Positive generators with gcd 1, each at most {MAX_GENERATOR}. "
-            "The presentation may be reordered or redundant; graph vertices use "
-            "its increasing minimal generator axis."
-        ),
-    )
-    value: CanonicalInteger
-
-    @model_validator(mode="after")
-    def require_complete_materialization(self) -> Self:
-        generators = _require_minimal_generators(self.generators)
-        value = _require_bounded_value(self.value)
-        _require_materializable_factorizations(
-            generators, value, MAX_GRAPH_FACTORIZATIONS
-        )
-        return self
-
-
-class FactorizationGraphComputeResult(StrictModel):
-    """Standard factorization graph with connected components."""
-
-    value: CanonicalInteger
-    minimal_generators: tuple[CanonicalInteger, ...] = Field(
-        min_length=1, max_length=MAX_GENERATORS
-    )
-    in_semigroup: bool
-    factorizations: tuple[tuple[int, ...], ...]
-    edges: tuple[tuple[int, int], ...]
-    connected_components: tuple[tuple[int, ...], ...]
-    is_connected: bool
-
-    @model_validator(mode="after")
-    def require_graph_partition(self) -> Self:
-        generators = _require_canonical_minimal_axis(self.minimal_generators)
-        value = parse_canonical_integer(self.value)
-        vertex_count = len(self.factorizations)
-        if self.in_semigroup != bool(self.factorizations):
-            raise _validation_error("membership must agree with graph vertices")
-        vertices = tuple(
-            index for component in self.connected_components for index in component
-        )
-        if tuple(sorted(vertices)) != tuple(range(vertex_count)):
-            raise _validation_error(
-                "connected components must partition all graph vertices"
-            )
-        if self.is_connected != (len(self.connected_components) <= 1):
-            raise _validation_error("is_connected must agree with connected components")
-        for left, right in self.edges:
-            if not 0 <= left < right < vertex_count:
-                raise _validation_error("graph edge has invalid vertex indices")
-        _require_exact_factorization_family(generators, value, self.factorizations)
-        expected_edges, expected_components = _factorization_graph_data(
-            self.factorizations
-        )
-        if self.edges != expected_edges:
-            raise _validation_error("edges do not match shared-support adjacency")
-        if self.connected_components != expected_components:
-            raise _validation_error("connected components do not match the graph")
         return self
 
 
@@ -1102,18 +830,14 @@ __all__ = [
     "ElementDeltaSetResult",
     "ElementElasticityRequest",
     "ElementElasticityResult",
-    "FactorizationComputeRequest",
-    "FactorizationComputeResult",
-    "FactorizationDistanceRequest",
-    "FactorizationDistanceResult",
-    "FactorizationGraphComputeRequest",
-    "FactorizationGraphComputeResult",
-    "FactorizationLengthsComputeRequest",
-    "FactorizationLengthsComputeResult",
     "MinimalPresentationRelation",
     "MinimalPresentationRequest",
     "MinimalPresentationResult",
+    "NumericalSemigroupSummaryRequest",
+    "NumericalSemigroupSummaryResult",
     "PresentationBinomial",
     "PresentationBinomialsRequest",
     "PresentationBinomialsResult",
+    "SemigroupMembershipRequest",
+    "SemigroupMembershipResult",
 ]
