+++ tests/math/numerical_semigroups/test_correctness_regressions.py
@@ -7,6 +7,10 @@ from math import lcm
 
 from tests.math.numerical_semigroups._support import numerical_semigroup_error
 
+from jacobian.math.numerical_semigroups._factorizations import (
+    FactorizationComputeRequest,
+    FactorizationGraphComputeRequest,
+)
 from jacobian.math.numerical_semigroups._models import (
     BettiElementsRequest,
     CatenaryDegreeRequest,
@@ -15,8 +19,6 @@ from jacobian.math.numerical_semigroups._models import (
     ElementCatenaryDegreeRequest,
     ElementDeltaSetRequest,
     ElementElasticityRequest,
-    FactorizationComputeRequest,
-    FactorizationGraphComputeRequest,
     MinimalPresentationRequest,
     MinimalPresentationResult,
     PresentationBinomialsRequest,
